Отрефакторит модуль со вспомогательными функциями.
