+++ src/utils/date.js
@@ -11,18 +11,6 @@ const TIME_IN_DAY_HOUR_MIN = 'DD[d] HH[h] mm[m]';
 const COUNT_OF_MS_IN_DAY = 86400000;
 const COUNT_OF_MS_IN_HOUR = 3600000;
 
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
-
-function getRandomInteger(a, b) {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-}
-
 function humanizeDate (date, format) {
   return date ? dayjs(date).format(format) : '';
 }
@@ -41,17 +29,10 @@ function getDuration (dateFrom, dateTo) {
   return dayjs.duration(difference).format(format);
 }
 
-function insertDashIntoStr (str) {
-  return str.trim().toLowerCase().replace(/[^a-zA-Z0-9 -]/, '').replace(/\s/g, '-');
-}
-
 export {
   DATE_FORMAT,
   TIME_FORMAT,
   DATE_TIME_FORMAT,
-  getRandomArrayElement,
-  getRandomInteger,
   humanizeDate,
-  getDuration,
-  insertDashIntoStr
+  getDuration
 };
